Split MCP workflow regression tests
